requirements.txt
ipykernel
ragatouille
python-dotenv
langchain
langchain-community
langchain-core
langchain-openai
langchain-pinecone
langchain-cohere
pydantic>=2.7.0
pinecone-client
pinecone-text>=0.8.0
jieba>=0.42.1
beautifulsoup4
tiktoken
numpy
langchainhub
chromadb
youtube-transcript-api<1.0.0
pytube
cohere
pypdf
yt_dlp